% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.r
\name{complement}
\alias{complement}
\title{compute DNA nucleotide complement}
\usage{
complement(nuc)
}
\arguments{
\item{input}{the character string of info to parse}
}
\description{
computes DNA complement
}
\examples{
anthrax_sasp.nuc
}
\keyword{parse}
\keyword{split,}
