% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nuc.r
\name{read_nuc}
\alias{read_nuc}
\title{read nucleotide}
\usage{
read_nuc(input_seq = NULL, file = NULL)
}
\arguments{
\item{input}{the string of character of info to parse}
}
\description{
parses nucleotide text of fasta file in text
}
\examples{
anthrax_sasp.nuc
}
\keyword{parse}
\keyword{split,}
