% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.r
\name{reverse_comp}
\alias{reverse_comp}
\title{reverse nucleotide}
\usage{
reverse_comp(seq)
}
\arguments{
\item{input}{the complement output to parse}
}
\description{
computes DNA reverse complement
}
\examples{
anthrax_sasp.nuc
}
\keyword{parse}
\keyword{split,}
