% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDNA.r
\name{read_nuc}
\alias{read_nuc}
\title{DNA revcomp}
\usage{
read_nuc(input_seq)
}
\arguments{
\item{input}{the vector of info to parse}

\item{split}{the delimiter}
}
\description{
computes DNA reverse complement
}
\examples{
myDNA()
}
\keyword{parse}
\keyword{split,}
